+++ plugins/plugin-chart-status/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/nielsen-oss/superset-viz-plugins/compare/@superset-viz-plugins/plugin-chart-status@0.1.1...@superset-viz-plugins/plugin-chart-status@0.1.2) (2021-01-19)
+
+**Note:** Version bump only for package @superset-viz-plugins/plugin-chart-status
+
+
+
+
+
+## 0.1.1 (2021-01-19)
+
+**Note:** Version bump only for package @superset-viz-plugins/plugin-chart-status
+
+
+
+
+
 ## [0.0.7](https://github.com/nielsen-oss/superset-viz-plugins/compare/@superset-viz-plugins/plugin-chart-status@0.0.6...@superset-viz-plugins/plugin-chart-status@0.0.7) (2021-01-04)
 
 **Note:** Version bump only for package @superset-viz-plugins/plugin-chart-status